Merge Staging into Prod

+++ scripts/rename-order-search-flag-to-internal.sh
@@ -14,6 +14,9 @@ ENV=$1
 
 CURRENT_INTERNAL_ORDER_SEARCH_VALUE=$(aws dynamodb get-item --region us-east-1 --table-name "efcms-deploy-${ENV}" --key '{"pk":{"S":"order-search-enabled"},"sk":{"S":"order-search-enabled"}}' | jq -r ".Item.current.BOOL")
 
-aws dynamodb put-item --region us-east-1 --table-name "efcms-deploy-${ENV}" --item '{"pk":{"S":"internal-order-search-enabled"},"sk":{"S":"internal-order-search-enabled"},"current":{"BOOL": '${CURRENT_INTERNAL_ORDER_SEARCH_VALUE}'}}'
+if [ ! -z "${CURRENT_INTERNAL_ORDER_SEARCH_VALUE}" ]; then
+  echo "deleting old order-search-enabled flag"
+  aws dynamodb delete-item --region us-east-1 --table-name "efcms-deploy-${ENV}" --key '{"pk":{"S":"order-search-enabled"},"sk":{"S":"order-search-enabled"}}'
+  aws dynamodb put-item --region us-east-1 --table-name "efcms-deploy-${ENV}" --item '{"pk":{"S":"internal-order-search-enabled"},"sk":{"S":"internal-order-search-enabled"},"current":{"BOOL": '${CURRENT_INTERNAL_ORDER_SEARCH_VALUE}'}}'
+fi
 
-aws dynamodb delete-item --region us-east-1 --table-name "efcms-deploy-${ENV}" --key '{"pk":{"S":"order-search-enabled"},"sk":{"S":"order-search-enabled"}}'
